--- src/main/java/za/ac/cput/Address.java
@@ -1,4 +1,0 @@
-package za.ac.cput;
-
-public class Address {
-}
