--- src/main/java/za/ac/cput/City.java
@@ -1,4 +1,0 @@
-package za.ac.cput;
-
-public class City {
-}
