--- src/main/java/za/ac/cput/Student.java
@@ -1,4 +1,0 @@
-package za.ac.cput;
-
-public class Student {
-}
