+++ src/main/java/za/ac/cput/domain/Address.java
@@ -1,0 +1,4 @@
+package za.ac.cput.domain;
+
+public class Address {
+}
