+++ src/main/java/za/ac/cput/domain/Country.java
@@ -1,0 +1,4 @@
+package za.ac.cput.domain;
+
+public class Country {
+}
