+++ src/main/java/za/ac/cput/domain/Name.java
@@ -1,0 +1,5 @@
+package za.ac.cput.domain;
+
+public class Name {
+
+}
